[Blueprints] Separate v2 runtime configuration resolution
